Fixed skid NParticle read

+++ pycosie/cluster/skid.py
@@ -62,7 +62,7 @@ class SkidCatalog():
                 self.pos = pos * L
                 self.ids = ID.astype(int)
                 self.velocity = ds.arr(velocity * unitVel.value * np.sqrt(a), "cm/s")
-                self.nParticle = int(npart)
+                self.nParticle = npart.astype(int)
                 self.vcmax = ds.arr(vcmax * unitVel * a, "cm/s") # maybe FIXME
                 self.hvc = ds.arr(hvc * unitVel * a, "cm/s") # maybe FIXME
                 self.dv = ds.arr(dv * unitVel * a, "cm/s") # maybe FIXME
